+++ main.cpp
@@ -11,8 +11,8 @@ int main() {
 
     int s=2; //ilość słów binarnych o długości 4 bity
     int w=4; //długość słowa - 4 bity
-    unsigned short a[2] = {6,2}; //(0110 0010)
-    unsigned short b[2] = {5,3}; //(0101 0011)
+    unsigned int a[2] = {6,2}; //(0110 0010)
+    unsigned int b[2] = {5,3}; //(0101 0011)
 
     Counter counter;
